Refactor dashboard to home, add learn page, update routing
